Move Linux & MacOs CallArranger tests into separate files
